requirements.txt
# Core dependencies for Claude API learning
anthropic>=0.25.0
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0

# Data science and visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Environment and utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Optional: for advanced experiments
streamlit>=1.28.0
gradio>=4.0.0